Support overriding metadata server with GCE_METADATA_HOST
This environment variable is used in all of the standard clients of the
google metadata service that I know of, and allows us to use a custom
metadata server.

In our case, we use this on non-GCE machines that have a custom method
of acquiring service account tokens. Instead of updating every tool to
support this custom method, we have a small metadata server running on
localhost that provides a subset of the standard metadata interface.

Example implementation in the python google auth library:
https://github.com/googleapis/google-auth-library-python/blob/18c2ec1b571d506c0dbcffc483aa5e7b95e1b246/google/auth/compute_engine/_metadata.py#L36

Change-Id: I7ff2e98d75c109c1027f3767a7efbf16eca783e9